embedding/requirements.txt
fastapi==0.116.1
torch==2.5.1
transformers==4.46.3
uvicorn==0.29.0
pillow==10.4.0
python-multipart==0.0.9